Fixed rendering errors from MC upgrade

+++ src/main/java/uristqwerty/gui_craftguide/texture/TintedTexture.java
@@ -14,17 +14,17 @@ public class TintedTexture implements Texture
 	@TextureParameter
 	public Color color;
 
-	private double prev[] = new double[4];
+	private float prev[] = new float[4];
 
 	@Override
 	public void renderRect(RendererBase renderer, int x, int y, int width, int height, int u, int v)
 	{
 		renderer.getColorModifierv(prev);
 		renderer.setColorModifier(
-				prev[0] * (color.red / 255.0),
-				prev[1] * (color.green / 255.0),
-				prev[2] * (color.blue / 255.0),
-				prev[3] * (color.alpha / 255.0));
+				prev[0] * (color.red / 255.0f),
+				prev[1] * (color.green / 255.0f),
+				prev[2] * (color.blue / 255.0f),
+				prev[3] * (color.alpha / 255.0f));
 
 		source.renderRect(renderer, x, y, width, height, u, v);
 		renderer.setColorModifierv(prev);
